feat: Update loaders to allow customization of size and colors
This commit modifies the RotatingSquaresLoader, ImageSpinLoader, and DNALoader components to allow customization of the size and colors. The size prop is now used to set the width and height of the loaders, and the squareColors prop is used to define the colors of the rotating squares in the RotatingSquaresLoader component. These changes enhance the flexibility and visual appearance of the loaders.